tentacle/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
requests==2.31.0
httpx==0.28.1
pyyaml==6.0.2
python-multipart==0.0.12
aiofiles==24.1.0
apscheduler==3.10.4
Pillow==10.4.0